Provide a flag to control generation of the invocation comment
The invocation comment stores the complete details of all provided
arguments, which may include paths to files. These aren't always
consistent, which results in unwanted changes in the generated files
(especially for CI checking for unexpected changes...).

Since the invocation isn't always useful, provide a flag to control
whether the comment is generated. To preserve current behaviour this
defaults to true (enabled).

Signed-off-by: Stephen Kitt <skitt@redhat.com>