Add --write_command_comment=false option (#153)
The invocation command can be user specific, especially if absolute
paths are used for the destination. This in turn can cause very large
needless diffs that will create lots of source control noise and be
distracting in PRs.

To remove these command comments simply pass
`--write_command_comment=false` on the CLI. The default behavior is
maintained.

Fixes #104.