Make Cond Matcher generic (#183)
`gomock.Cond` at the moment expects an matcher function with the
signature `func(x any) bool` requiring a manual cast to the expected
type. By making this matcher generic the callback function can be typed
removing the need for a manual cast to the expected type.

The change should be backwards compatible with existing tests using
`gomock.Cond`.