chore: upgrade go versions and fix ci test execution (#187)
I updated the CI to use the two newest stable versions of go (v1.21.x
and v1.22.x) to test the code as well as the oldest version (`v1.19.x`)
that could still be used to build the source.

The behaviour of the `go version` line in `go.mod` changed with
`v1.21.x`: it should state the oldest version of go that is needed to
build the source. The source builds and tests fine with `v1.19.x`, older
versions however fail, so I set it to this version.

Before my changes the `test` GH workflow didn't actually execute
`./ci/test.sh` because there is was check that skipped testing if the
used go version was any other than `1.18`. I removed this check which
showed some tests failing even without my changes.

The issue turned out to have been introduced by accident (because tests
weren't executed) in https://github.com/uber-go/mock/pull/48. The fix
for it can be found in commit
https://github.com/uber-go/mock/pull/187/commits/0403497b4ea38317a8cbf0fb30c122f4dc60bdb4
(part of this PR). Additionally I had to regenerate all mocks or
`./ci/tests.sh` would complain that they weren't up to date.