Clarify documentation on Finish
Finish no longer has to be explicitly called on Controller with defer
if the Controller was created with testing.T.

We tried to just mark this as deprecated in #50 but doing so would break
many users, so we reverted it in #85.

In doing so we dropped some wordings from the doc that marked Finished
as unnecessary for most cases.

This PR fixes the docstring to guide users without explicitly marking
the API deprecated.