Add gob mode to support custom package loading techniques in place of `--exec_only`
Change #207 replaced reflect mode with package mode.
In doing so, we no longer had to create separate reflect programs
that would encode the resulting package into a gob
for the rest of mockgen to read from.
Because of that, we removed the `--exec_only` flag.

However, this flag was actually used to pass in programs that created gob encodings
in other ways, allowing customization of the package loading and encoding portion of mockgen.
As an alternative, this adds a new flag to mockgen `--model_gob`,
which allows mockgen users to pass a path to a gob encoding of a `model.Package`,
so that custom ways of loading package information can still be supported.